Move stuff out of Library.tsx
changes in Tabcomponent
